+++ R/new-data.R
@@ -3,7 +3,9 @@
 #' \lifecycle{stable}
 #'
 #' @param .data A `tbl_ts`.
-#' @param n An integer indicates the number of key-index pair to append.
+#' @param n An integer indicates the number of key-index pair to append. If
+#' * `n > 0`, future observations
+#' * `n < 0`, past observations
 #' @param ... Passed to individual S3 method.
 #'
 #' @rdname new-data
@@ -20,9 +22,10 @@ new_data <- function(.data, n = 1L, ...) {
 #' new_data(pedestrian)
 #' new_data(pedestrian, keep_all = TRUE)
 #' new_data(pedestrian, n = 3)
+#' new_data(pedestrian, n = -2)
 new_data.tbl_ts <- function(.data, n = 1L, keep_all = FALSE, ...) {
-  if (!is_integerish(n, 1) && any(n > 0)) {
-    abort("Argument `n` must be a positive integer.")
+  if (!is_integerish(n, 1)) {
+    abort("Argument `n` must be an integer.")
   }
   abort_if_irregular(.data)
   abort_unknown_interval(int <- interval(.data))
@@ -32,7 +35,14 @@ new_data.tbl_ts <- function(.data, n = 1L, keep_all = FALSE, ...) {
 
   key_data <- key_data(.data)
   grped_df <- new_grouped_df(.data, groups = key_data)
-  last_entry <- summarise(grped_df, !!idx := max(!!idx))
+  if (n >= 0) {
+    is_ord <- TRUE
+    last_entry <- summarise(grped_df, !!idx := max(!!idx))
+  } else {
+    is_ord <- NULL
+    last_entry <- summarise(grped_df, !!idx := min(!!idx))
+    tunit <- -tunit
+  }
   if (NCOL(key_data) == 1) { # no key
     regrped_df <- last_entry
   } else {
@@ -40,7 +50,7 @@ new_data.tbl_ts <- function(.data, n = 1L, keep_all = FALSE, ...) {
     regrped_df <- new_grouped_df(last_entry, groups = meta_grps)
   }
   new_lst <- mutate(regrped_df, 
-    !!idx := list2(tibble(!!idx := seq(!!idx, by = tunit, length.out = n + 1)[-1])))
+    !!idx := list2(tibble(!!idx := seq(!!idx, by = tunit, length.out = abs(n) + 1)[-1])))
 
   out <- unwrap(ungroup(new_lst), .col = !!idx)
   if (keep_all) {
@@ -48,17 +58,18 @@ new_data.tbl_ts <- function(.data, n = 1L, keep_all = FALSE, ...) {
   } else { # reorder column names according to the data input
     out <- out[setdiff(names(.data), measured_vars(.data))]
   }
-  update_meta(out, .data, ordered = TRUE, interval = interval(.data))
+  update_meta(out, .data, ordered = is_ord, interval = interval(.data))
 }
 
 #' @description
-#' `append_row()`: add new rows to the end of a tsibble by filling a key-index
+#' `append_row()`: add new rows to the start/end of a tsibble by filling a key-index
 #' pair and `NA` for measured variables.
 #'
 #' `append_case()` is an alias of `append_row()`.
 #' @rdname new-data
 #' @export
 #' @examples
+#'
 #' tsbl <- tsibble(
 #'   date = rep(as.Date("2017-01-01") + 0:2, each = 2),
 #'   group = rep(letters[1:2], 3),
@@ -67,6 +78,7 @@ new_data.tbl_ts <- function(.data, n = 1L, keep_all = FALSE, ...) {
 #' )
 #' append_row(tsbl)
 #' append_row(tsbl, n = 2)
+#' append_row(tsbl, n = -2)
 append_row <- function(.data, n = 1L, ...) {
   UseMethod("append_row")
 }